refactor: create recommendation.utils for recommendation functions
